refactor: Voucher 인터페이스를 추상 클래스로 변경
