+++ src/main/java/team/marco/vouchermanagementsystem/VoucherManagementSystemApplication.java
@@ -5,7 +5,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
-import team.marco.vouchermanagementsystem.model.VoucherType;
 
 @SpringBootApplication
 public class VoucherManagementSystemApplication {
