--- src/main/java/team/marco/vouchermanagementsystem/model/Voucher.java
@@ -1,11 +1,0 @@
-package team.marco.vouchermanagementsystem.model;
-
-import java.util.UUID;
-
-public interface Voucher {
-    UUID getId();
-
-    String getInfo();
-
-    VoucherType getType();
-}
