--- src/main/java/team/marco/vouchermanagementsystem/model/VoucherType.java
@@ -1,5 +1,0 @@
-package team.marco.vouchermanagementsystem.model;
-
-public enum VoucherType {
-    FIXED, PERCENT
-}
