+++ src/main/java/team/marco/vouchermanagementsystem/model/voucher/FixedAmountVoucher.java
@@ -1,16 +1,18 @@
-package team.marco.vouchermanagementsystem.model;
+package team.marco.vouchermanagementsystem.model.voucher;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.text.MessageFormat;
 import java.util.UUID;
 
 import static java.text.MessageFormat.format;
+import static team.marco.vouchermanagementsystem.model.voucher.VoucherType.*;
 
-public class FixedAmountVoucher implements Voucher {
+public class FixedAmountVoucher extends Voucher {
     private static final Logger logger = LoggerFactory.getLogger(FixedAmountVoucher.class);
 
-    private final UUID id;
+    private final VoucherType type = FIXED;
     private final int amount;
 
     public FixedAmountVoucher(int amount) {
@@ -18,33 +20,28 @@ public class FixedAmountVoucher implements Voucher {
             throw new IllegalArgumentException(format("{0}: 할인 금액은 0 보다 커야 합니다.", amount));
         }
 
-        this.id = UUID.randomUUID();
         this.amount = amount;
 
-        logger.debug("Create FixedAmountVoucher {id: {}, amount: {}}", id, amount);
+        logger.debug("Create FixedAmountVoucher {id: {}, amount: {}}", getId(), amount);
     }
 
     public FixedAmountVoucher(UUID id, int amount) {
-        this.id = id;
+        super(id);
+
         this.amount = amount;
-    }
-
-    @Override
-    public UUID getId() {
-        return id;
-    }
-
-    @Override
-    public String getInfo() {
-        return format("{0}원 할인 쿠폰", amount);
     }
 
     @Override
     public VoucherType getType() {
-        return VoucherType.FIXED;
+        return type;
     }
 
     public int getAmount() {
         return amount;
+    }
+
+    @Override
+    public String toString() {
+        return MessageFormat.format("FixedAmountVoucher'{'type={0}, amount={1}'}'", type, amount);
     }
 }
