+++ src/main/java/team/marco/vouchermanagementsystem/model/voucher/Voucher.java
@@ -1,0 +1,21 @@
+package team.marco.vouchermanagementsystem.model.voucher;
+
+import java.util.UUID;
+
+public abstract class Voucher {
+    private UUID id;
+
+    public Voucher() {
+        this.id = UUID.randomUUID();
+    }
+
+    public Voucher(UUID id) {
+        this.id = id;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    abstract public VoucherType getType();
+}
