+++ src/main/java/team/marco/vouchermanagementsystem/model/voucher/VoucherType.java
@@ -1,0 +1,5 @@
+package team.marco.vouchermanagementsystem.model.voucher;
+
+public enum VoucherType {
+    FIXED, PERCENT
+}
