+++ src/main/java/team/marco/vouchermanagementsystem/repository/MemoryVoucherRepository.java
@@ -2,7 +2,7 @@ package team.marco.vouchermanagementsystem.repository;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
-import team.marco.vouchermanagementsystem.model.Voucher;
+import team.marco.vouchermanagementsystem.model.voucher.Voucher;
 
 import java.util.HashMap;
 import java.util.List;
